[Lodgify.Cinema.Infrastructure.Data] namespace refactor
